The SCREEN_PING_OF_DEATH check was dead code:
  if (meta->pkt_len > 65535) ...

pkt_len was __u16, so the comparison was always false and
GLOBAL_CTR_SCREEN_PING_OF_DEATH never incremented. Same dead code
in xdp_screen.c and tc_screen_egress.c.

Widen to __u32. With u32, the check is now meaningful — fires for
oversized frames (jumbo / GSO segmented) that legitimately exceed
65535 bytes. The classical ping-of-death attack requires fragment
reassembly to overflow 65535; xpf doesn't reassemble, so this is
not a complete reimplementation, but the comparison is no longer
trivially-false dead code.

Also fixes the IPv6 wrap noted in the issue:

  meta->pkt_len = bpf_ntohs(ip6h->payload_len) + 40;
  meta->pkt_len = bpf_ntohs(ip6h->payload_len) + sizeof(struct ipv6hdr);

bpf_ntohs returns u16; the addition wrapped u16 even when the
target was u32. Cast to u32 first.

Closes #860.

+++ bpf/headers/xpf_common.h
@@ -440,7 +440,12 @@ struct pkt_meta {
 	__u16 l3_offset;
 	__u16 l4_offset;
 	__u16 payload_offset;
-	__u16 pkt_len;
+	/* #860: pkt_len is u32, not u16. IPv4 tot_len fits in u16, but
+	 * IPv6 payload_len + sizeof(ipv6hdr) can wrap u16 for jumbo
+	 * frames (payload > 65495). Widening also makes the
+	 * SCREEN_PING_OF_DEATH `pkt_len > 65535` check meaningful —
+	 * with u16 the comparison was always false (dead code). */
+	__u32 pkt_len;
 
 	/* Zone classification */
 	__u16 ingress_zone;
